+++ committee/db/mysql/util/doc.go
@@ -1,2 +1,0 @@
-// Package util is used to avoid circular dependencies between packages.
-package util