+++ contrib/screener-api/screener/suite_test.go
@@ -80,7 +80,7 @@ func (s *ScreenerSuite) TestScreener() {
 	s.T().Setenv("TRM_URL", "")
 
 	cfg := config.Config{
-		AppSecret: "appsecret",
+		AppSecret: "secret",
 		AppID:     "appid",
 		TRMKey:    "",
 		Rulesets: map[string]config.RulesetConfig{
@@ -155,7 +155,7 @@ func (s *ScreenerSuite) TestScreener() {
 
 	// now test crud screener
 	blacklistBody := client.BlackListBody{
-		TypeReq: "create",
+		Type:    "create",
 		ID:      "1",
 		Data:    "{\"test\":\"data\"}",
 		Address: "0x123",
@@ -166,27 +166,31 @@ func (s *ScreenerSuite) TestScreener() {
 
 	// post to the blacklist
 	status, err := apiClient.BlacklistAddress(s.GetTestContext(), cfg.AppSecret, cfg.AppID, blacklistBody)
+	fmt.Println(status)
 	Equal(s.T(), "success", status)
 	Nil(s.T(), err)
 
 	// update an address on the blacklist
-	blacklistBody.TypeReq = "update"
+	blacklistBody.Type = "update"
 	blacklistBody.Remark = "new remark"
 
 	status, err = apiClient.BlacklistAddress(s.GetTestContext(), cfg.AppSecret, cfg.AppID, blacklistBody)
+	fmt.Println(status)
 	Equal(s.T(), "success", status)
 	Nil(s.T(), err)
 
 	// delete the address on the blacklist
-	blacklistBody.TypeReq = "delete"
+	blacklistBody.Type = "delete"
 	blacklistBody.ID = "1"
 
 	status, err = apiClient.BlacklistAddress(s.GetTestContext(), cfg.AppSecret, cfg.AppID, blacklistBody)
+	fmt.Println(status)
 	Equal(s.T(), "success", status)
 	Nil(s.T(), err)
 
 	// unauthorized
 	status, err = apiClient.BlacklistAddress(s.GetTestContext(), "bad", cfg.AppID, blacklistBody)
+	fmt.Println(status)
 	NotEqual(s.T(), "success", status)
 	NotNil(s.T(), err)
 }