+++ packages/synapse-interface/constants/bridge.ts
@@ -44,32 +44,6 @@ export const BRIDGE_CONFIG_ADDRESSES = {
   [CHAINS.POLYGON.id]: '0xd69229f223a8fc84998e1361ae7b4ff724cf4a49', // TESTING ADDRESS
 }
 
-/**
- * number of required confirmations from bridge
- */
-export const BRIDGE_REQUIRED_CONFIRMATIONS = {
-  [CHAINS.ETH.id]: 33,
-  [CHAINS.BNB.id]: 33,
-  [CHAINS.POLYGON.id]: 128,
-  [CHAINS.FANTOM.id]: 80,
-  [CHAINS.BOBA.id]: 33,
-  [CHAINS.OPTIMISM.id]: 750,
-  [CHAINS.MOONBEAM.id]: 33,
-  [CHAINS.MOONRIVER.id]: 33,
-  [CHAINS.ARBITRUM.id]: 200,
-  [CHAINS.AVALANCHE.id]: 80,
-  [CHAINS.DFK.id]: 33,
-  [CHAINS.HARMONY.id]: 33,
-  [CHAINS.AURORA.id]: 33,
-  [CHAINS.CRONOS.id]: 33,
-  [CHAINS.METIS.id]: 33,
-  [CHAINS.DOGE.id]: 33,
-  [CHAINS.CANTO.id]: 20,
-  [CHAINS.BASE.id]: 750,
-  [CHAINS.KLAYTN.id]: 20,
-  [CHAINS.BLAST.id]: 750,
-}
-
 export const DEFAULT_FROM_TOKEN_SYMBOL = 'USDC'
 export const DEFAULT_TO_TOKEN_SYMBOL = 'USDC'
 export const DEFAULT_FROM_TOKEN = USDC