+++ sin-executor/contracts/executionservice/generate.go
@@ -1,3 +1,0 @@
-package executionservice
-
-//go:generate go run github.com/synapsecns/sanguine/tools/abigen generate --sol ../../../packages/contracts-communication/flattened/SynapseExecutionServiceV1Harness.sol --pkg executionservice --sol-version 0.8.20 --filename executionservice --evm-version istanbul